fix #182: introduce RtAsyncType to handle "Promise" return type
Converting interfaces that declare async methods via return type
`Task` fails currently. The result does not contain any Promise
return type, but just `any`. A return value of `Task` is an
implicit declaration of asynchronous function. Since
interfaces to not contain implementation of functions, the
explicit defining keyword `async` is not supported.

So far, automatic detection of async functions was insufficient.

This commit introduces a new return type `RtAsyncType` that
handles proper creation of TypeScript type declaration.
Once a method returns `Task` and ÀutoAsync` flag is enabled,
the return value `Promise` is used with TypeScript code.

If the  `AutoAsync` flag is not enabled, then the async
 (Promise) value is unfolded instead of just converted to
`any`.

Converting an empty function body takes the "async" keyword
into consideration, too. If a function is NOT "async" but
returns a Promise, then a proper Promise is created instead
of just `null`.