detect nullable environment on C# >= 8.0
With C# compiler higher than 8.0 a new pragma flag
was introduced to invert the nullability of reference types.
If this flag is enabled, reference types are not implicitly
nullable anymore. If the nullability is being explicitly set,
then such types should be exported with this nullability
to TypeScript too.

If this flag is not set, then the new code will take no effect.

see:
* https://github.com/dotnet/roslyn/blob/master/docs/features/nullable-metadata.md
* https://codeblog.jonskeet.uk/2019/02/10/nullableattribute-and-c-8/