fix: export as interface with third-party base class
If a class-type is exported as an interface but derives
from a third-party base class, a type cast exception was thrown.
The exported expected a class-node (`RtClass`) during the export
but apparently it is a `RtInterface`, because the type is
being exported as interface.

So, base classes are added to the list of `implementees` instead
as there is no base class for an interface.

fixes: https://github.com/reinforced/Reinforced.Typings/issues/189