fix: introduce `FieldExportBuilder`
A field export builder is being introduced to provide
the properly typed `Member` property. Fields use
`FieldInfo`, whereas properties use `PropertyInfo`.
So, using original `PropertyExportBuilder` with
accessing the member info leads to a type cast
exception.

Since the new `FieldExportBuilder` is a descendant
of `PropertyExportBuilder`, it can be used as a
replacement seemlesly - where needed.

fixes:
https://github.com/reinforced/Reinforced.Typings/issues/190