fix: omit parameter initialiser in interfaces
Methods in interface may not have parameter default values
("initialisers"). TypeScript compiler 3.9.7 emits error TS2371
if such a default value is being found.

Such default value makes a parameter optional. So, instead
of emitting the default value, the parameter is marked with "?".
Then in can be omitted from the list of parameters as it
would be with a default value.

fixes https://github.com/reinforced/Reinforced.Typings/issues/191